refactor(loop): extract floor-gate + fix ANSI leak in guardrail markdown
Two fixes the self-guardrail flagged on the release PR:

1. The correctness-floor wiring pushed stop-gate.ts over the 500-line large-file
   threshold, a net-new finding. Move the loop's floor integration
   (buildFloorGate, captureFloorSnapshot, the entry-snapshot diff, ledger-status
   mapping, repair message) into src/loop/floor-gate.ts — a cohesive unit —
   leaving stop-gate.ts focused on the guardrail/ledger flow and back under the
   threshold (478 lines). No behavior change; imports re-pointed in cli.ts and
   the stop-gate test.

2. `guardrail check --markdown` still wrote the ANSI-styled console header +
   "Checking…" + baseline lines to stdout before the markdown, so escape codes
   leaked into the machine-captured report (piped to a file / posted as a PR
   comment). Suppress those console lines in --markdown as well as --json (a new
   `quiet` guard) so the markdown report is pure markdown.
